requirements.txt
streamlit>=1.30
requests>=2.30
beautifulsoup4>=4.12
Pillow>=10.0
cairosvg>=2.7
qrcode[pil]>=7.4
